Remove unused settings, move lightning to wallet
